requirements.txt
google-generativeai
groq
openai
ollama
colorama